Make downloadKite method of adapters
It makes it easier to add special cases for a given platform at the cost
of some duplication between osx and windows.
Another advantage is that it will make the creation of a test adapter
much easier as all operations requiring to hit the network can now
completely avoided in that adapter.